requirements.txt
python-telegram-bot==20.6
google-generativeai
requests
python-dotenv
PyPDF2
